Start of merge of Sampler classes.
